+++ src/main/java/com/hospital/mateus/curso/medico/dto/DadosCadastroMedico.java
@@ -9,6 +9,7 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import org.hibernate.validator.constraints.br.CPF;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 public record DadosCadastroMedico(
@@ -28,7 +29,7 @@ public record DadosCadastroMedico(
 
         @NotNull
         @Positive
-        Double salario,
+        BigDecimal salario,
 
         @NotBlank
         String crm,
